+++ circle.yml
@@ -1,4 +1,4 @@
-machine:
+clearmachine:
   environment:
     IMPORT_PATH: "github.com/$CIRCLE_PROJECT_USERNAME/$CIRCLE_PROJECT_REPONAME"
