+++ cmd/actuary/testdata/auditctl2/auditctl
@@ -1,0 +1,2 @@
+#!/bin/sh
+echo "some data that fails"
