+++ cmd/actuary/testdata/docker.service
@@ -1,0 +1,2 @@
+#!/bin/sh
+echo "docker.service"
