+++ cmd/actuary/testdata/docker.socket
@@ -1,0 +1,2 @@
+#!/bin/sh
+echo "docker.socket"
