+++ cmd/actuary/testdata/etc/docker/certs.d/certFolder/cert
@@ -1,0 +1,2 @@
+#!/bin/sh
+echo "cert"
