+++ cmd/actuary/testdata/etc/docker/daemon.json
@@ -1,0 +1,2 @@
+#!/bin/sh
+echo "docker.socket"
