+++ cmd/actuary/testdata/fstabFail
@@ -1,0 +1,1 @@
+/filler /wrong
