+++ cmd/actuary/testdata/fstabPass
@@ -1,0 +1,1 @@
+/filler /var/lib/docker
