+++ cmd/actuary/testdata/groupPass
@@ -1,0 +1,1 @@
+docker:users: user1, user2, user3
