+++ cmd/actuary/testdata/var/run/docker.sock
@@ -1,0 +1,2 @@
+#!/bin/sh
+echo "docker.sock"
